check loginsession to render to public or private website

+++ project.py
@@ -177,7 +177,10 @@ def restaurantsJSON():
 @app.route('/restaurant/')
 def showRestaurants():
   	restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-  	return render_template('restaurants.html', restaurants = restaurants)
+  	if 'username' not in login_session:
+  		return render_template('publicrestaurant.html', restaurants = restaurants)
+  	else:
+  		return render_template('restaurants.html', restaurants = restaurants)
 
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
@@ -213,6 +216,12 @@ def editRestaurant(restaurant_id):
 def deleteRestaurant(restaurant_id):
 
   	restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  	if 'username' not in login_session:
+  		return redirect('/login')
+  	if restaurantToDelete.user_id != login_session['user_id']:
+  		return "<script> function myFunction() {alert('You are not authorized to delete\
+  		this restaurant. Please create your own restaurant in order to delete.');}</script>\
+  		<body onload = 'myFunction()'> "
   	if request.method == 'POST':
 		session.delete(restaurantToDelete)
 		flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -225,10 +234,13 @@ def deleteRestaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/')
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
-
 	restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+	creator = getUserInfo(restaurant.user_id)
 	items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-	return render_template('menu.html', items = items, restaurant = restaurant)
+	if 'username' not in login_session or creator.id != login_session['user_id']:
+		return render_template('publicmenu.html', items = items, restaurant = restaurant, creator = creator)
+	else:
+		return render_template('menu.html', items = items, restaurant = restaurant,creator = creator)
